requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
celery[redis]==5.2.7
flower==1.2.0
celery-singleton==0.3.1
django-redis==5.2.0
django-cachalot==2.5.2
environs==9.5.0
django-filter==23.2
social-auth-app-django==5.3.0
